Grouping: The other devices too.
