+++ sw/lib/include/neorv32_mtime.h
@@ -43,6 +43,23 @@
 #ifndef neorv32_mtime_h
 #define neorv32_mtime_h
 
+/**********************************************************************//**
+ * @name IO Device: Machine System Timer (MTIME)
+ **************************************************************************/
+/**@{*/
+/** MTIME module prototype */
+typedef volatile struct __attribute__((packed,aligned(4))) {
+  uint32_t TIME_LO;    /**< offset 0:  time register low word */
+  uint32_t TIME_HI;    /**< offset 4:  time register high word */
+  uint32_t TIMECMP_LO; /**< offset 8:  compare register low word */
+  uint32_t TIMECMP_HI; /**< offset 12: compare register high word */
+} neorv32_mtime_t;
+
+/** MTIME module hardware access (#neorv32_mtime_t) */
+#define NEORV32_MTIME ((neorv32_mtime_t*) (NEORV32_MTIME_BASE))
+/**@}*/
+
+
 // prototypes
 int      neorv32_mtime_available(void);
 void     neorv32_mtime_set_time(uint64_t time);
